This adds the ability to take in pre-sentenced text.
There were some nasty subtleties here.
Specifically, we need to take into account both inter-section and
inter-sentence white-space in the original text. If the original
document is (white-space important):

[{Foo bar.} {Foo baz.}]  [{Bar baz.}]

where [] represents Sections and {} represents Sentences, then we
need to be sure to account for the single whitespace between the two
sentences in the first section. We also need to account for the two
whitespace characters between the two sections. Note that this is
different from

[{Foo bar.} {Foo baz.}  ][{Bar baz.}]

or

[{Foo bar. }{Foo baz.}]  [{Bar baz.}]

or

[{Foo bar. }{Foo baz. } ][{Bar baz.}]

since in each of these three alternative cases, CoreNLP records any
whitespace before and after tokens. We were properly using that recorded
information, but CoreNLP can only record what we give it.