This fixes #28.
"Non-processed" sections (those that only have tokenization and
sentence-splitting run) are now properly represented in annotated
Communications. Previously, the `Section` was being transferred
over (due to the PerspectiveCommunication), and as such, the
`rawTextSpan` was set. However, the `textSpan` was not set, and neither
were any `Sentences` (or `Tokenizations`). This has now been fixed.